Re-introduce epochs to re-frame-multi integration
The regular re-frame integration has the concept of "epochs". The idea
is that you can discard "stale" events related to a state that was
started in a prior epoch. See
https://lucywang000.github.io/clj-statecharts/docs/integration/re-frame/#discard-stale-events-with-epoch-support
for a more detailed explanation.

This commit brings the idea of epochs to the re-frame-multi integration,
but in a somewhat simpler implementation. Now the machine as a whole has
an epoch. When it initializes a state, it assigns its current epoch to
the state. There is a event handler that advances a machine's epoch.
Then any transition events on states that were created in an earlier
epoch are ignored.