Structure scheduler to interoperate with data stores
There are a few situations where this library is asked to store state
as it changes over time, including services and the re-frame
integration. Notably, these are the two situations which permit use of a
scheduler.

A scheduler has two purposes. First, it must provide an interface to
schedule a transition at a future time. Second, when that transition
runs, it must be able to affect its environment somehow. That is, it
must store or transmit the new state somewhere. In the case of a
service, the transitioned state is saved within the service and passed
on to any listeners. In the case of the re-frame integration, the state
is persisted in re-frame's app-db. It is expected that any other
integrations will follow a similar pattern: create a fsm that uses a
scheduler to schedule transitions and to save/transmit the transitioned
state somewhere.

This commit introduces the idea of a data store: the destination
environment that is updated by transitions, whether those transitions
are triggered internally by the scheduler or externally by an app.

This provides a common interface for a stateful environment (a service
or integration) and a scheduler to collaborate around.

It also breaks a circular dependency between them. Since the environment
and scheduler can share a reference to a store, they no longer need
references to each other.